Ported magic candles from SLASH'EM.
For a while I've had an evil idea in the back of my mind. That idea was to replace the magic lamps at Izchak's with magic candles to reduce the average wishes per game (WPG). This commit ports the candles and does exactly that. I also restored the probability of the magic light source to it's vanilla prob (3 vs 2).

There were a couple of extras in HackEM but I did not bother to port those, only the essentials for this item to function correctly.