Fix: extracting gold dragon scales from minvent didn't call end_burn
From xNetHack commit c70d6bcdf:

'This happened because the condition for end_burn made the (unchanged
from vanilla) assumption that the only light source in a monster's
inventory that would be worn would be body armor, not a cloak. Since
gold dragon scales are a cloak in xNetHack but still provide light, the
condition needed to be expanded to include that. Otherwise, the light
source will stick around and wind up either panicking or crashing the
game.'