Fix: mirrors were not providing reflection after being shattered.
Huge oversight in my initial implementation of mirrors providing slotless reflection. When the mirror was shattered, it would correctly reduce the damage dice of the zap that was affected, but as for the actual zap effects they would ignore reflection. I discovered this thankfully from a zapped card of death which revealed how bad this bug was.

My fix is slightly clunky, but as is my standard it should be readable and easy to understand for others.