Be more generous with hydration.
After playtesting a bit of grung, I found that the early game was pretty brutal. Any grung was pretty much forced to blitz for the Sokoban rivers ASAP, and we don't really want to force that for every game... So I eased up on the hydration for a lot of effects like fountain quaffing.