Increase speed and decrease AC of spheres.
This change is influenced by c536603 from EvilHack. In that commit, the spell-beings for the flame sphere and freeze sphere spells were fully converted to separate 'orb' monsters. They mostly remained the same as their sphere counterparts but their speed was increased to 15 and AC was increased to 10. I like the idea of the change, but I think the KISS route is just changing sphere stats across the board. This kind of buffs the Orb Weaver monster but also makes their orbs easier to hit.