New artifact: The Glyph Shard.
After a lot of playtesting, I've noticed that some games require extensive backtracking if you don't get a decent setup of items (ie: bag of holding or teleport control). However, there's also been a couple ideas from other variants in the back of my mind, notably magic chests and thiefstones. Both of these items address the "problem" of backtracking in different ways. In NerfHack style, I want to attempt my own approach without using either of those solutions.

My idea is a kind of portal stone. This stone can be invoked to teleport to a specific level that you have keyed in. To prevent this from being too overpowered (and possibly rendering teleport scrolls moot), it can only be keyed once and the level it's keyed to is set permanently for the rest of the game. For something like your stash level, which might be quite early in the dungeon or somewhere like Sokoban, this could be extremely useful.

At first, I toyed with the idea of a new type of gray stone, but ultimately I decided to make this a unique artifact, so that it's powers are kept on a leash. I tossed it into the Wyrm Caves so that players have a good reason to ransack that level (once they can take on the dragons of course).

Hopefully this will reward players for reaching the roughly-halfway point of the game and streamline things.