Fix: Greasy feet were not wearing off. Fixes #135.
Getting fumbling from a grease trap would incorrectly set the timeout for the fumbling to 0, causing the timeout to never trigger and making the player permanently greased. Very unfortunate. I noticed I didn't use longs where I should have, probably resulting in the 0 init value.

I believe this also fixed 135 because if the value was 0 it makes sense that when applying a towel the check for fumbling might not evaluate correctly.