Fix: potions and forge dipping (from EvilHack).
In EvilHack it was noticed that there was no effect when potions were dipped into forges. I tested this in Nerf and it was also the case. However, most other items that the Evil commit addressed like scrolls, spellbooks, and globs already seemed to be handled. I just had to update lava_damage so that potions were added and vapor effects took effect if the player was close.