Buff crossbow bolts.
The crossbow is a fairly MEH launcher. Since limiting it's multishot to zero, my intention was for it to be much stronger, but that only starts happening at skilled or higher. We need to beef it up at basic for it to be even remotely considerable as a weapon. Also, since the undead slayer can start with one, it should be desireable for them to want to use it instead of just start scumming for a silver weapon or just tossing them in favor of daggers.

I raised both the damage vs small and large monsters. The increase for small goes from 1d4+1 to 1d7+1. The increase for large monsters is still from 1d6+1 to 1d8+1.

I also increased the to-hit bonus from +0 to +2. I think this makes sense since with a crossbow you can actually aim it. Compare this to daggers which always get a +2 bonus, these should arguably get +3 but that seems a bit high.