Fix: Re-frozen ice in Sokoban could melt.
This is something that we really don't want in case the ice was in a place critical to solving a level. Before, if something melted the ice, we could zap it with a cold ray and re-freeze it. However, a melt-timer would be added, meaning that the ice would eventually revert to a pool again. Not good.