Fix: monsters would not travel over puddles.
I noticed this in the orcish mine's end where there is a circle of puddles and a mordor orc just stood in the middle not moving an inch.

My fix here is not perfect but it does solve the problem. I noticed that giant eels don't strictly stick to puddles but also will hop on land briefly. For now this behavior seems ok, but maybe we'll come back to it. The monster moving code is a little tricky to navigate and I don't want to prod it further...